+++ splunk-otel-android/src/main/java/com/splunk/rum/ZipkinToDiskSender.java
@@ -6,7 +6,6 @@ import java.io.File;
 import java.io.IOException;
 import java.time.Clock;
 import java.util.List;
-import java.util.Locale;
 
 import zipkin2.Call;
 import zipkin2.codec.Encoding;
@@ -17,16 +16,13 @@ class ZipkinToDiskSender extends Sender {
     private final File path;
     private final FileUtils fileUtils;
     private final Clock clock;
+    private final DeviceSpanStorageLimiter storageLimiter;
 
-    ZipkinToDiskSender(File path) {
-        this(path, new FileUtils(), Clock.systemDefaultZone());
-    }
-
-    // exists for testing
-    ZipkinToDiskSender(File path, FileUtils fileUtils, Clock clock) {
-        this.path = path;
-        this.fileUtils = fileUtils;
-        this.clock = clock;
+    private ZipkinToDiskSender(Builder builder) {
+        this.path = builder.path;
+        this.fileUtils = builder.fileUtils;
+        this.clock = builder.clock;
+        this.storageLimiter = builder.storageLimiter;
     }
 
     @Override
@@ -46,6 +42,10 @@ class ZipkinToDiskSender extends Sender {
 
     @Override
     public Call<Void> sendSpans(List<byte[]> encodedSpans) {
+        if (!storageLimiter.ensureFreeSpace()) {
+            Log.e(SplunkRum.LOG_TAG, "Dropping " + encodedSpans.size() + " spans: Too much telemetry has been buffered or not enough space on device.");
+            return Call.create(null);
+        }
         long now = clock.millis();
         File filename = createFilename(now);
         try {
@@ -58,5 +58,40 @@ class ZipkinToDiskSender extends Sender {
 
     private File createFilename(long now) {
         return new File(path, now + ".spans");
+    }
+
+    static Builder builder() {
+        return new Builder();
+    }
+
+    static class Builder {
+        private File path;
+        private FileUtils fileUtils = new FileUtils();
+        private Clock clock = Clock.systemDefaultZone();
+        private DeviceSpanStorageLimiter storageLimiter;
+
+        Builder path(File path){
+            this.path = path;
+            return this;
+        }
+
+        Builder fileUtils(FileUtils fileUtils){
+            this.fileUtils = fileUtils;
+            return this;
+        }
+
+        Builder clock(Clock clock){
+            this.clock = clock;
+            return this;
+        }
+
+        Builder storageLimiter(DeviceSpanStorageLimiter limiter){
+            this.storageLimiter = limiter;
+            return this;
+        }
+
+        ZipkinToDiskSender build() {
+            return new ZipkinToDiskSender(this);
+        }
     }
 }